+++ prisma/migrations/20250213163240_removed/migration.sql
@@ -1,0 +1,2 @@
+-- DropIndex
+DROP INDEX "Folder_accountId_key";